requirements.txt
litellm>=1.40.0
datasets>=2.19.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
tqdm>=4.66.0
python-dotenv>=1.0.0
